Factor out all_starting_or_stopping_containers logic.
Surgery: factor out the logic of figuring out which one of mutually
dependent containers can be actioned first. This separates the logic
that defines which containers are eligible to be started, and the
fact that the containers are being started, from state-independent
logic of taking the defined relationships between the objects and
working out which ones have no dependencies any more and can be
started next.