Add separate before/after dependencies.
requires and required_by are now limited in scope of pulling in which
containers to start and stop together.

Once the containers are in starting or stopping state it's the before/after
dependencies that govern their relative order. Autogenerate before/after
from requires and required by, to avoid requiring them explicitly.

Also: pull in all starting and stopping dependencies together into
a single call to do_start() and do_stop(), and have it work on all of them
at once. This would be the case most of the time, and this actually seems
simpler, no need to figure out all_starting_or_stopping containers together.