inittab conversion fixes
Fixes dependency issues when switching between levels 3/4/5 directly.
Remove the extra_stopping_command_for_started units, that stops rc.M
services. Instead of running "stop rc.M.target" in the -started units,
run it as the stop command of the system/inittab/rc unit, which remains
started when switching the runlevels. This fixes rc.M services getting
stopped in error, during this switch.

Drop the nowait flag, and fix race the race condition it was needed for,
instead of leaving it up on the user to know to use --nowait when running
vera commands as part of a unit. Redirect stdout to the initial vlad
command's stdout, and any nested vlad commands will not redirect the stdout
again. Use a weak object to track active stdout redirect.

Clean up the piggybacking of start requests, decided to not attach to
a running start, if attempting to start the a unit that's already starting,
this also removes the need to hold multiple requesters.